+++ components/HomeScreen.js
@@ -1,0 +1,25 @@
+import React from "react";
+import { StyleSheet, Text, View, Button } from "react-native";
+
+export default class HomeScreen extends React.Component {
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text>Welcome to Stamp Rally!</Text>
+        <Button
+          title="Open details!"
+          onPress={() => this.props.navigation.navigate("Details")}
+        />
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center"
+  }
+});
